Created popup modal for auth sections.

+++ src/app/web/shows/single-show/show-navigation/show-navigation.component.ts
@@ -1,4 +1,8 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
+import { Subscription } from 'rxjs/Subscription';
+
+import { AuthService } from '../../../auth/auth.service';
+import { ModalService } from '../../../../shared/services/modal.service';
 import { ProfileService } from '../../../profile/profile.service';
 
 @Component({
@@ -6,19 +10,36 @@ import { ProfileService } from '../../../profile/profile.service';
   templateUrl: './show-navigation.component.html',
   styleUrls: ['./show-navigation.component.scss']
 })
-export class ShowNavigationComponent implements OnInit {
+export class ShowNavigationComponent implements OnInit, OnDestroy {
   @Input() showId: number;
   @Input() isWatched: boolean = false;
   @Input() isFavourite: boolean = false;
   @Input() rating: string = 'N/A';
+  isUserAuthenticated: boolean = false;
+  subscription: Subscription;
 
-  constructor(private profileService: ProfileService) {
+  constructor(
+    private profileService: ProfileService,
+    private authService: AuthService,
+    private modalService: ModalService) {
   }
 
   ngOnInit() {
+    this.isUserAuthenticated = this.authService.isUserLoggedIn();
+    this.subscription = this.authService.userAuthenticationChanged
+      .subscribe(
+        (user: any) => {
+          this.isUserAuthenticated = !!user;
+        }
+      );
   }
 
   toggleFavourite() {
+    if (!this.isUserAuthenticated) {
+      this.modalService.open('sign-in');
+      return;
+    }
+
     if (this.isFavourite) {
       this.profileService.removeShowFromFavourites(this.showId).subscribe((res: any) => {
         this.isFavourite = res.result;
@@ -31,6 +52,11 @@ export class ShowNavigationComponent implements OnInit {
   }
 
   toggleWatched() {
+    if (!this.isUserAuthenticated) {
+      this.modalService.open('sign-in');
+      return;
+    }
+
     if (this.isWatched) {
       this.profileService.removeShowToWatchlist(this.showId).subscribe((res: any) => {
         this.isWatched = res.result;
@@ -40,5 +66,9 @@ export class ShowNavigationComponent implements OnInit {
         this.isWatched = res.result;
       });
     }
+  }
+
+  ngOnDestroy() {
+    this.subscription.unsubscribe();
   }
 }
